requirements.txt
aioredis==1.2.0
appdirs==1.4.3
asgiref==3.1.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
black==19.3b0
channels==2.1.7
channels-redis==2.4.0
Click==7.0
constantly==15.1.0
daphne==2.3.0
Django==2.1.7
django-cors-headers==2.5.2
django-webpack-loader==0.5.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.1
msgpack-python==0.5.6
psycopg2==2.8.2
PyHamcrest==1.9.0
pytz==2019.1
redis==3.2.1
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
Twisted==19.2.0
txaio==18.8.1
zope.interface==4.6.0
